+++ src/model.rs
@@ -198,7 +198,7 @@ impl Model {
     pub(super) fn to_model_db_entry(
         &mut self,
         timestamp: f64,
-        deck_id: usize,
+        deck_id: i64,
     ) -> Result<ModelDbEntry, Error> {
         self.templates
             .iter_mut()
@@ -233,7 +233,7 @@ impl Model {
     }
 
     #[allow(dead_code)]
-    pub(super) fn to_json(&mut self, timestamp: f64, deck_id: usize) -> Result<String, Error> {
+    pub(super) fn to_json(&mut self, timestamp: f64, deck_id: i64) -> Result<String, Error> {
         Ok(
             serde_json::to_string(&self.to_model_db_entry(timestamp, deck_id)?)
                 .map_err(json_error)?,
